Updates sweeper logging to use standard `tf_resource_type` field
