+++ internal/sweep/context.go
@@ -6,8 +6,6 @@ package sweep
 import (
 	"context"
 
-	"github.com/hashicorp/go-hclog"
-	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-log/tfsdklog"
 )
 
@@ -17,17 +15,6 @@ func Context(region string) context.Context {
 	ctx = tfsdklog.RegisterStdlogSink(ctx)
 
 	ctx = logger(ctx, "sweeper", region)
-
-	return ctx
-}
-
-func logger(ctx context.Context, loggerName, region string) context.Context {
-	ctx = tfsdklog.NewRootProviderLogger(ctx,
-		tfsdklog.WithLevel(hclog.Debug),
-		tfsdklog.WithLogName(loggerName),
-		tfsdklog.WithoutLocation(),
-	)
-	ctx = tflog.SetField(ctx, "sweeper_region", region)
 
 	return ctx
 }
